renamed cloud to integrations
